engine/scene_generator/requirements.txt
# Core dependencies
numpy>=1.21.0
torch>=2.0.0
Pillow>=9.0.0
tqdm>=4.65.0

# Noise generation
noise>=1.2.2

# Image processing and generation
diffusers>=0.19.0
transformers>=4.30.0
accelerate>=0.20.0

# Semantic matching
sentence-transformers>=2.2.2

# Configuration and serialization
pyyaml>=6.0
jsonschema>=4.17.0

# Utilities
scipy>=1.10.0
scikit-image>=0.20.0

# Optional: GPU support
# cuda-python>=12.0.0  # Uncomment if using CUDA 